# Datetime performance - `vec_ptype2()`