--- a/src/utils.c
+++ b/src/utils.c
@@ -15,8 +15,10 @@ SEXP strings_tbl = NULL;
 SEXP strings_tbl_df = NULL;
 SEXP strings_data_frame = NULL;
 SEXP strings_vctrs_rcrd = NULL;
-SEXP strings_posixt = NULL;
+SEXP strings_date = NULL;
+SEXP strings_posixct = NULL;
 SEXP strings_posixlt = NULL;
+SEXP strings_posixt = NULL;
 SEXP strings_factor = NULL;
 SEXP strings_ordered = NULL;
 SEXP strings_vctrs_vctr = NULL;
@@ -1184,7 +1186,7 @@ void vctrs_init_utils(SEXP ns) {
 
   // Holds the CHARSXP objects because unlike symbols they can be
   // garbage collected
-  strings = Rf_allocVector(STRSXP, 18);
+  strings = Rf_allocVector(STRSXP, 20);
   R_PreserveObject(strings);
 
   strings_dots = Rf_mkChar("...");
@@ -1196,50 +1198,56 @@ void vctrs_init_utils(SEXP ns) {
   strings_vctrs_rcrd = Rf_mkChar("vctrs_rcrd");
   SET_STRING_ELT(strings, 2, strings_vctrs_rcrd);
 
+  strings_date = Rf_mkChar("Date");
+  SET_STRING_ELT(strings, 3, strings_date);
+
+  strings_posixct = Rf_mkChar("POSIXct");
+  SET_STRING_ELT(strings, 4, strings_posixct);
+
   strings_posixlt = Rf_mkChar("POSIXlt");
-  SET_STRING_ELT(strings, 3, strings_posixlt);
+  SET_STRING_ELT(strings, 5, strings_posixlt);
 
   strings_posixt = Rf_mkChar("POSIXt");
-  SET_STRING_ELT(strings, 4, strings_posixlt);
+  SET_STRING_ELT(strings, 6, strings_posixlt);
 
   strings_vctrs_vctr = Rf_mkChar("vctrs_vctr");
-  SET_STRING_ELT(strings, 5, strings_vctrs_vctr);
+  SET_STRING_ELT(strings, 7, strings_vctrs_vctr);
 
   strings_none = Rf_mkChar("none");
-  SET_STRING_ELT(strings, 6, strings_none);
+  SET_STRING_ELT(strings, 8, strings_none);
 
   strings_minimal = Rf_mkChar("minimal");
-  SET_STRING_ELT(strings, 7, strings_minimal);
+  SET_STRING_ELT(strings, 9, strings_minimal);
 
   strings_unique = Rf_mkChar("unique");
-  SET_STRING_ELT(strings, 8, strings_unique);
+  SET_STRING_ELT(strings, 10, strings_unique);
 
   strings_universal = Rf_mkChar("universal");
-  SET_STRING_ELT(strings, 9, strings_universal);
+  SET_STRING_ELT(strings, 11, strings_universal);
 
   strings_check_unique = Rf_mkChar("check_unique");
-  SET_STRING_ELT(strings, 10, strings_check_unique);
+  SET_STRING_ELT(strings, 12, strings_check_unique);
 
   strings_key = Rf_mkChar("key");
-  SET_STRING_ELT(strings, 11, strings_key);
+  SET_STRING_ELT(strings, 13, strings_key);
 
   strings_loc = Rf_mkChar("loc");
-  SET_STRING_ELT(strings, 12, strings_loc);
+  SET_STRING_ELT(strings, 14, strings_loc);
 
   strings_val = Rf_mkChar("val");
-  SET_STRING_ELT(strings, 13, strings_val);
+  SET_STRING_ELT(strings, 15, strings_val);
 
   strings_group = Rf_mkChar("group");
-  SET_STRING_ELT(strings, 14, strings_group);
+  SET_STRING_ELT(strings, 16, strings_group);
 
   strings_length = Rf_mkChar("length");
-  SET_STRING_ELT(strings, 15, strings_length);
+  SET_STRING_ELT(strings, 17, strings_length);
 
   strings_factor = Rf_mkChar("factor");
-  SET_STRING_ELT(strings, 16, strings_factor);
+  SET_STRING_ELT(strings, 18, strings_factor);
 
   strings_ordered = Rf_mkChar("ordered");
-  SET_STRING_ELT(strings, 17, strings_ordered);
+  SET_STRING_ELT(strings, 19, strings_ordered);
 
 
   classes_data_frame = Rf_allocVector(STRSXP, 1);