--- a/src/utils.h
+++ b/src/utils.h
@@ -272,6 +272,8 @@ extern SEXP strings_tbl;
 extern SEXP strings_tbl_df;
 extern SEXP strings_data_frame;
 extern SEXP strings_vctrs_rcrd;
+extern SEXP strings_date;
+extern SEXP strings_posixct;
 extern SEXP strings_posixlt;
 extern SEXP strings_posixt;
 extern SEXP strings_factor;