--- a/src/utils.h
+++ b/src/utils.h
@@ -19,6 +19,9 @@ enum vctrs_class_type {
   vctrs_class_bare_factor,
   vctrs_class_bare_ordered,
   vctrs_class_rcrd,
+  vctrs_class_bare_date,
+  vctrs_class_bare_posixct,
+  vctrs_class_bare_posixlt,
   vctrs_class_posixlt,
   vctrs_class_unknown,
   vctrs_class_none