--- a/tests/testthat/test-type-date-time.R
+++ b/tests/testthat/test-type-date-time.R
@@ -93,6 +93,15 @@ test_that("POSIXlt always steered towards POSIXct", {
   expect_equal(vec_ptype2(dtl, dtl), dtc[0])
 })
 
+test_that("vec_ptype2() on a POSIXlt with multiple time zones returns the first", {
+  rlang::with_options(TZ = "America/New_York", {
+    x <- as.POSIXlt(new_datetime(tzone = ""))
+  })
+
+  expect_identical(attr(x, "tzone"), c("", "EST", "EDT"))
+  expect_identical(attr(vec_ptype2(x, new_date()), "tzone"), "")
+})
+
 test_that("vec_ptype2(<Date>, NA) is symmetric (#687)", {
   date <- new_date()
   expect_identical(