This PR adds native C level `vec_ptype2()` dispatch for `Date`, `POSIXct` and `POSIXlt` classes

This was fairly straightforward now that we have the dispatch infrastructure in place. I've separated `vctrs_class_bare_posixct` from `vctrs_class_bare_posixlt` because I'm fairly certain it will be useful to have the distinction when we go to implement the cast methods. For the ptype2 methods it doesn't matter as much.

We get nice performance improvements all around.

``` r
library(vctrs)

date <- new_date()
datetime <- new_datetime(tzone = "UTC")
```

Two dates, nothing to do

```r
# before
bench::mark(vec_ptype2(date, date))
#> # A tibble: 1 x 6
#>   expression                  min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>             <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(date, date)   15.2µs   17.7µs    54523.    6.66KB     81.9

# after
bench::mark(vec_ptype2(date, date))
#> # A tibble: 1 x 6
#>   expression                  min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>             <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(date, date)    842ns   1.23µs   793154.        0B     159.
```

Two datetimes, takes the "union" of the tzones

```r
# before
bench::mark(vec_ptype2(datetime, datetime))
#> # A tibble: 1 x 6
#>   expression                          min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                     <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(datetime, datetime)   19.3µs   22.5µs    40507.    25.1KB     20.3

# after
bench::mark(vec_ptype2(datetime, datetime))
#> # A tibble: 1 x 6
#>   expression                          min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                     <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(datetime, datetime)   1.23µs   1.41µs   516363.        0B        0
```

Date and datetime, uses tzone of the datetime

```r
# before
bench::mark(vec_ptype2(datetime, date))
#> # A tibble: 1 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(datetime, date)   16.5µs   18.4µs    51397.    4.38KB     25.7

# after
bench::mark(vec_ptype2(datetime, date))
#> # A tibble: 1 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(datetime, date)   1.23µs   1.45µs   676618.        0B        0
```